07-langchain-framework/requirements.txt
ipykernel
langchain
langchain-community
langchain-openai
openai
huggingface_hub
transformers
jinja2
tiktoken
pypdf
yt_dlp
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pydub
